Use 64 bit hash

+++ imgui.h
@@ -230,7 +230,7 @@ typedef unsigned short ImDrawIdx;   // Default: 16-bit (for maximum compatibilit
 #endif
 
 // Scalar data types
-typedef unsigned int        ImGuiID;// A unique ID used by widgets (typically the result of hashing a stack of string)
+typedef unsigned long long  ImGuiID;// A unique ID used by widgets (typically the result of hashing a stack of string)
 typedef signed char         ImS8;   // 8-bit signed integer
 typedef unsigned char       ImU8;   // 8-bit unsigned integer
 typedef signed short        ImS16;  // 16-bit signed integer
